requirements.txt
blinker==1.6.2
click==8.1.3
dnspython==2.3.0
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
Pillow==9.5.0
pymongo==4.3.3
pytesseract==0.3.10
python-decouple==3.8
Werkzeug==2.3.4
